More adjustments
Add Keyboard.pause and Keyboard.resume to do a soft disconnect/reconnect of the keyboard. This is useful, for instance, to pause/resume after detecting the Wootility as even so much as having a HID connection to the LED/Configuration interface of the keyboard causes the Wootility to freak out.
Remove a couple unused variables in LedController and AnalogController.
Fix PreciseTimer in layer system to more immediately end the timer in end().